% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_packages.R
\name{assert_all_programs_exist_in_path}
\alias{assert_all_programs_exist_in_path}
\title{Assert programs are in PATH}
\usage{
assert_all_programs_exist_in_path(
  x,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{Program name(s) (character vector)}

\item{msg}{A character string containing the error message to display if one
or more programs are missing from PATH}

\item{call}{Only relevant when pooling assertions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if all programs in \code{x} are available in PATH,
otherwise aborts with the message specified by \code{msg}
}
\description{
Assert that one or more executable programs are available on the system PATH.
}
\examples{
rscript <- if (.Platform$OS.type == "windows") "Rscript.exe" else "Rscript"

try({
  assert_all_programs_exist_in_path(rscript) # Passes
  assert_all_programs_exist_in_path(c(rscript, "this_program_should_not_exist_123"))
})

}
