% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_create_advanced}
\alias{assert_create_advanced}
\title{Create an advanced assertion function}
\usage{
assert_create_advanced(func)
}
\arguments{
\item{func}{A function defining the assertion criteria.
This function should return a TRUE when assertion should pass,
or a character string representing an error message if it fail. E
Error message strings can contain the same special keywords / formatting as described in \code{\link[=assert_create]{assert_create()}}}
}
\value{
Returns a custom assertion function that can be used to check if a supplied value meets the defined criteria.
}
\description{
Creates a custom assertion function that can be used to check if a supplied value meets certain criteria.
The input function is required to return TRUE if the assertion should pass or a string error message if it should fail.
}
\concept{assert_create}
