% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_type.R
\name{assert_factor}
\alias{assert_factor}
\title{Assert that the input object is a factor}
\usage{
assert_factor(x, msg = NULL, call = rlang::caller_env())
}
\arguments{
\item{x}{An object}

\item{msg}{A character string containing the error message to display if \code{x} is not a factor}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}
}
\value{
invisible(TRUE) if \code{x} is a factor, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert that the input object is a factor
}
\examples{
\dontrun{
assert_factor(factor(c("a", "b", "c"))) # Passes
assert_factor(c("a", "b", "c")) # Throws default error
assert_factor(factor(c("a", "b", "c")), "Custom error message") # Passes
assert_factor(1:3, "Custom error message") # Throws custom error
}

}
