% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_functions.R
\name{assert_function_expects}
\alias{assert_function_expects}
\title{Assert function expects required argument names}
\usage{
assert_function_expects(
  x,
  required,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{a function to check includes required argument names}

\item{required}{a character vector of required argument names}

\item{msg}{The error message thrown if the assertion fails (string)}

\item{call}{Only relevant when pooling assertions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if function \code{x} expects all required arguments, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert a function expects a set of required argument names.
}
