% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_compare.R
\name{assert_greater_than_or_equal_to}
\alias{assert_greater_than_or_equal_to}
\title{Assert that the input object is greater than or equal to a specified minimum value}
\usage{
assert_greater_than_or_equal_to(
  x,
  minimum,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{An object to check}

\item{minimum}{The minimum value to compare against}

\item{msg}{A character string containing the error message to display if \code{x} is not greater than or equal to the specified minimum value (string)}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if \code{x} is greater than or equal to the specified minimum value, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert that the input object is greater than or equal to a specified minimum value
}
\examples{
try({
assert_greater_than_or_equal_to(3, 2) # Passes
assert_greater_than_or_equal_to(c(3, 4, 5), 2) # Passes
assert_greater_than_or_equal_to(2, 3) # Throws error
})
}
