% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{assert_has_no_missing_values}
\alias{assert_has_no_missing_values}
\title{Assert that the input vector has no missing values}
\usage{
assert_has_no_missing_values(
  x,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{A vector.}

\item{msg}{A character string containing the error message to display if \code{x} has missing values.}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if \code{x} has no missing values, otherwise aborts with the error message specified by \code{msg}
}
\description{
This function asserts that the input vector has no missing values (\code{NA}) and aborts
with an error message if it does.
}
\examples{
\dontrun{
assert_has_no_missing_values(c(1, 2, 3))  # Passes
assert_has_no_missing_values(c(1, NA, 2))  # Throws default error

assert_has_no_missing_values(c(1, 2, 3), msg = "Custom error message")  # Passes
assert_has_no_missing_values(c(1, NA, 2), msg = "Custom error message")  # Throws custom error
}

}
\concept{assert_has}
