% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_compare.R
\name{assert_less_than_or_equal_to}
\alias{assert_less_than_or_equal_to}
\title{Assert input is less than or equal to a specified maximum value}
\usage{
assert_less_than_or_equal_to(
  x,
  maximum,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{An object to check}

\item{maximum}{The maximum value to compare against}

\item{msg}{A character string containing the error message to display if \code{x} is not less than or equal to the specified maximum value (string)}

\item{call}{Only relevant when pooling assertions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if \code{x} is less than or equal to the specified maximum value, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert a number is less than or equal to a specified maximum value.
For vectorized version see \code{\link[=assert_all_less_than_or_equal_to]{assert_all_less_than_or_equal_to()}}
}
\examples{
try({
assert_less_than_or_equal_to(1, 2) # Passes
assert_less_than_or_equal_to(c(1, 2, 3), 3) # Throws error
assert_less_than_or_equal_to(3, 2) # Throws error
})
}
