% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_type.R
\name{assert_logical_vector}
\alias{assert_logical_vector}
\title{Assert that the input object is an atomic logical vector}
\usage{
assert_logical_vector(x, msg = NULL, call = rlang::caller_env(), ...)
}
\arguments{
\item{x}{An object}

\item{msg}{A character string containing the error message to display if x is not an atomic logical vector}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}
}
\value{
invisible(TRUE) if x is an atomic logical vector, otherwise aborts with the error message specified by msg
}
\description{
Assert that the input object is an atomic logical vector
}
\examples{
\dontrun{
assert_logical_vector(c(TRUE, TRUE, TRUE)) # Passes
assert_logical_vector("a") # Throws default error
assert_logical_vector(c(1, 0, 1), "Custom error message") # Throws custom error
}
}
