% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_type.R
\name{assert_numeric_vector}
\alias{assert_numeric_vector}
\title{Assert input is a numeric vector}
\usage{
assert_numeric_vector(
  x,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{An object}

\item{msg}{A character string containing the error message to display if \code{x} is not a numeric vector}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if \code{x} is a numeric vector, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert input is a numeric vector
}
\concept{assert_type}
