% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_packages.R
\name{assert_program_exists_in_path}
\alias{assert_program_exists_in_path}
\title{Assert program is in PATH}
\usage{
assert_program_exists_in_path(
  x,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{x}{Program name (string)}

\item{msg}{A character string containing the error message to display if the
program is missing from PATH}

\item{call}{Only relevant when pooling assertions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if \code{x} is available in PATH, otherwise aborts with
the message specified by \code{msg}
}
\description{
Assert that a single executable program is available on the system PATH.
To assert multiple programs, use \code{\link[=assert_all_programs_exist_in_path]{assert_all_programs_exist_in_path()}}.
}
\examples{
rscript <- if (.Platform$OS.type == "windows") "Rscript.exe" else "Rscript"

try({
  assert_program_exists_in_path(rscript) # Passes
  assert_program_exists_in_path("this_program_should_not_exist_123")
})

}
