% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_type.R
\name{assert_string}
\alias{assert_string}
\title{Assert that the input object is a character string}
\usage{
assert_string(x, msg = NULL, call = rlang::caller_env(), ...)
}
\arguments{
\item{x}{An object}

\item{msg}{A character string containing the error message to display if x is not a character vector}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}
}
\value{
invisible(TRUE) if x is a character vector, otherwise aborts with the error message specified by msg
}
\description{
Assert that the input object is a character string
}
\examples{
\dontrun{
assert_string("a") # Passes
assert_string(c("a", "b", "c")) # Throws default error
assert_string(1:3) # Throws default error
assert_string(c("a", 1, "b"), "Custom error message") # Throws custom error
}
}
