% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_testing.R
\name{assertion_tests}
\alias{assertion_tests}
\title{Count tests per Assertion}
\usage{
assertion_tests()
}
\value{
two column data.frame describing assertion name and number of tests (expect_statement)
}
\description{
Count the number of unit-tests per assertion.
Note assertion_tests only finds tests where \code{expect_} and \code{assert_} are on the same line.
}
