% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_testing.R
\name{check_all_assertions_are_tested_enough}
\alias{check_all_assertions_are_tested_enough}
\title{Check assertions are tested enough}
\usage{
check_all_assertions_are_tested_enough(min_required_tests = 5)
}
\arguments{
\item{min_required_tests}{min number of tests (expect statements) per assertion}
}
\value{
TRUE if all assertions sufficiently tested. Otherwise throws error
}
\description{
Check assertions are tested enough
}
