% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_roxygen_params}
\alias{common_roxygen_params}
\title{Common Parameter Descriptions}
\usage{
common_roxygen_params(call, arg_name, msg, ...)
}
\arguments{
\item{call}{Only relevant when pooling assertions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}

\item{msg}{The error message thrown if the assertion fails (string)}

\item{...}{Used to pass any arguments to assertion function}
}
\description{
Common Parameter Descriptions
}
