% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_regex.R
\name{contains_pattern}
\alias{contains_pattern}
\title{Check if strings contain a regex pattern}
\usage{
contains_pattern(
  x,
  pattern,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{x}{A character vector to check}

\item{pattern}{A regular expression pattern (string)}

\item{ignore.case, perl, fixed, useBytes}{Logical flags passed to \code{\link[=grepl]{grepl()}}}
}
\value{
TRUE if all strings in \code{x} match \code{pattern}, otherwise FALSE or a string error message.
}
\description{
This helper checks whether all elements of a character vector match a regex pattern.
}
\concept{assert_regex}
