% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{excludes_advanced}
\alias{excludes_advanced}
\title{Check if an object does not contain prohibited elements}
\usage{
excludes_advanced(x, illegal)
}
\arguments{
\item{x}{An object to check}

\item{illegal}{The prohibited elements to check for}
}
\value{
Returns TRUE if \code{x} is the same type as \code{illegal} and \code{x} does not include any of the \code{illegal} elements.
Otherwise returns a string representing the appropriate error message to display
}
\description{
This function checks that \code{x} does not include any of the \code{illegal} elements.
\code{x} must be the same type as \code{illegal}.
Factors are treated as character vectors.
}
