% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_functions.R
\name{function_expects_advanced}
\alias{function_expects_advanced}
\title{Expect function has arguments}
\usage{
function_expects_advanced(x, required)
}
\arguments{
\item{required}{the names of parameters the function must support (character)}
}
\value{
TRUE if the function contains the expected parameters, otherwise returns a string (the error message)
}
\description{
Expect function has arguments
}
