% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_names.R
\name{has_all_names}
\alias{has_all_names}
\title{Check if a named object has all specified names}
\usage{
has_all_names(x, names)
}
\arguments{
\item{x}{a named object}

\item{names}{A character vector of names to check for in \code{x}.}
}
\value{
A logical value indicating whether \code{x} has all the names specified in \code{names}.
}
\description{
This function returns a logical value indicating whether the object \code{x} has all the names specified in \code{names}.
}
\examples{
assertions:::has_all_names(list(a = 1, b = 2), c("a", "b"))  # returns TRUE
assertions:::has_all_names(list(a = 1, b = 2), c("a", "c"))  # returns FALSE
}
