% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_class}
\alias{has_class}
\title{Check object is some class}
\usage{
has_class(x, class)
}
\arguments{
\item{x}{A value to check.}

\item{class}{checks if \code{x} belongs to \code{class}. If multiple values of \code{class} are supplied, returns whether \code{x} belongs to any of them (character)}
}
\value{
A logical scalar indicating \code{x} belongs to \code{class}
}
\description{
This function checks whether object is a specific class
}
\examples{
if(interactive()) {
  has_class(1, "numeric") # TRUE
  has_class(1, "character") # FALSE
}
}
\concept{is_type}
