% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_files.R
\name{has_extension}
\alias{has_extension}
\title{Title}
\usage{
has_extension(x, extensions, compression = FALSE)
}
\arguments{
\item{x}{object to test}

\item{extensions}{valid extensions (character vector). Do not include the '.', e.g. supply \code{extensions = 'txt'} not \code{extensions = '.txt'}}

\item{compression}{should compression extension ‘.gz’, ‘.bz2’ or ‘.xz’ be removed first?}
}
\value{
TRUE if all x have valid extensions as supplied by \code{extensions} (flag)
}
\description{
Title
}
