% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_missing_values}
\alias{has_missing_values}
\title{Check if a vector has missing values}
\usage{
has_missing_values(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical value indicating whether the input vector contains missing values.
}
\description{
This function returns a logical value indicating whether the input vector contains
missing values (\code{NA}).
}
\examples{
if(interactive()){
has_missing_values(c(1, 2, 3))  # returns FALSE
has_missing_values(c(1, NA, 2))  # returns TRUE
}
}
