% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_no_duplicates}
\alias{has_no_duplicates}
\title{Check if a vector has no duplicates}
\usage{
has_no_duplicates(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical value indicating whether the input vector contains no duplicated elements.
}
\description{
This function returns a logical value indicating whether the input vector contains
no duplicated elements.
}
\examples{
has_no_duplicates(c(1, 2, 3))  # returns TRUE
has_no_duplicates(c(1, 2, 2))  # returns FALSE

}
