% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has_no_missing_values}
\alias{has_no_missing_values}
\title{Check if a vector has no missing values}
\usage{
has_no_missing_values(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical value indicating whether the input vector contains no missing values.
}
\description{
This function returns a logical value indicating whether the input vector contains
no missing values (\code{NA}).
}
\examples{
if(interactive()){
has_no_missing_values(c(1, 2, 3))  # returns TRUE
has_no_missing_values(c(1, NA, 2))  # returns FALSE
}
}
