% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_files.R
\name{has_permission}
\alias{has_permission}
\title{Check file permissions}
\usage{
has_permission(filepaths, permission = c("write", "execute", "read"))
}
\arguments{
\item{filepaths}{A character vector of file paths to check}

\item{permission}{A character string of permission to check for. One of c('write', 'execute', 'read')}
}
\value{
A logical value indicating whether all files have the specified permissions
}
\description{
Check if the specified filepaths have the specified permissions
}
\examples{
\dontrun{
has_permission(c("file1.txt", "file2.txt"), permission = "rw")
}
}
