% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{includes}
\alias{includes}
\title{Check if All Values in Required are in x}
\usage{
includes(x, required)
}
\arguments{
\item{x}{A vector of elements.}

\item{required}{A vector of elements to check for inclusion in \code{x}.}
}
\value{
A logical value indicating whether all elements of \code{required} are present in \code{x} (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Checks if all elements of \code{required} are present in \code{x}.
}
