% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{includes_advanced}
\alias{includes_advanced}
\title{Check if an object contains required elements}
\usage{
includes_advanced(x, required)
}
\arguments{
\item{x}{An object to check}

\item{required}{The required elements to check for}
}
\value{
Returns TRUE if \code{x} is the same type as \code{required} and \code{x} includes all the \code{required} elements.
Otherwise returns a string representing the appropriate error message to display
}
\description{
This function checks that \code{x} includes all of the \code{required} elements.
\code{x} must be the same type as \code{required}.
Factors are treated as character vectors.
}
