% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_character_vector}
\alias{is_character_vector}
\title{Check if an object is a character vector}
\usage{
is_character_vector(x, include_lists = FALSE)
}
\arguments{
\item{x}{An object to check.}

\item{include_lists}{A logical value indicating whether to include lists in the check. Defaults to \code{FALSE}.}
}
\value{
A logical value indicating whether \code{x} is a character vector.
}
\description{
Check if an object is a character vector
}
