% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_character_vector_or_glue}
\alias{is_character_vector_or_glue}
\title{Check if an object is a character vector}
\usage{
is_character_vector_or_glue(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a character vector or glue vector.
}
\description{
Differs from \code{\link[=is_character_vector]{is_character_vector()}} in that it permits glue character vectors to pass.
}
\concept{is_type}
