% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_flag}
\alias{is_flag}
\title{Check if a value is a logical flag}
\usage{
is_flag(x)
}
\arguments{
\item{x}{A value to check.}
}
\value{
A logical scalar indicating whether \code{x} is a logical flag.
}
\description{
This function checks if a value is a logical scalar (i.e., a single logical value).
}
\examples{
is_flag(TRUE)
is_flag(FALSE)
is_flag(c(TRUE, FALSE))
}
\concept{is_type}
