% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_comparisons.R
\name{is_identical}
\alias{is_identical}
\title{Check if two objects are identical}
\usage{
is_identical(x, y)
}
\arguments{
\item{x}{first object to compare}

\item{y}{second object to compare}
}
\value{
logical value indicating whether or not the objects are identical
}
\description{
Check if two objects are identical
}
