% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_logical_vector}
\alias{is_logical_vector}
\title{Check if an object is a logical vector}
\usage{
is_logical_vector(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a logical vector.
}
\description{
Check if an object is a logical vector
}
\concept{is_type}
