% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_number}
\alias{is_number}
\title{Check if an object is a single number}
\usage{
is_number(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a single number.
}
\description{
Check if an object is a single number
}
