% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_numeric_vector}
\alias{is_numeric_vector}
\title{Check if an object is a numeric vector}
\usage{
is_numeric_vector(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a numeric vector.
}
\description{
This function checks if an object is a numeric vector in R.
}
\examples{
is_numeric_vector(c(1, 2, 3)) # TRUE
is_numeric_vector(list(1, 2, 3)) # FALSE
is_numeric_vector(1:5) # TRUE
is_numeric_vector("hello") # FALSE
is_numeric_vector(list(1, 2, "a")) # FALSE

}
\concept{is_type}
