% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_reactive}
\alias{is_reactive}
\title{Check if a value is reactive}
\usage{
is_reactive(x)
}
\arguments{
\item{x}{A value to check.}
}
\value{
A logical scalar indicating whether \code{x} is a list.
}
\description{
This function checks if a value is reactive
}
\examples{
\dontrun{
is_reactive(shiny::reactive(1)) # TRUE
is_reactive(1) # FALSE
}
}
\concept{is_type}
