% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_comparisons.R
\name{is_same_type}
\alias{is_same_type}
\title{Check equality of type}
\usage{
is_same_type(x, y)
}
\arguments{
\item{x}{first object to compare}

\item{y}{second object to compare}
}
\value{
TRUE if x and y are of the same type, otherwise FALSE
}
\description{
Is type of \code{x} the same as \code{y} (according to typof)
}
