% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_string}
\alias{is_string}
\title{Check if an object is a single string}
\usage{
is_string(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a single string.
}
\description{
Check if an object is a single string
}
\concept{is_type}
