% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_string_advanced}
\alias{is_string_advanced}
\title{Check if x is a string}
\usage{
is_string_advanced(x)
}
\arguments{
\item{x}{A value to be checked}
}
\value{
Returns invisible(TRUE) if x is a character value with length 1. Returns a string
with an error message if x is not a character value or has a length other than 1.
}
\description{
This function is designed for use with \code{assert_create_advanced}. It must return
TRUE for the assertion to pass or a string representing the error message if the
assertion should fail.
}
