% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{is_subset}
\alias{is_subset}
\title{Check if one set is a subset of another}
\usage{
is_subset(x, y)
}
\arguments{
\item{x}{A numeric, character, or logical vector.}

\item{y}{A numeric, character, or logical vector.}
}
\value{
A logical value indicating whether \code{x} is a subset of \code{y}.
}
\description{
Determines if all elements in set \code{x} are also present in set \code{y}.
}
