% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{is_superset}
\alias{is_superset}
\title{Check if one set is a superset of another}
\usage{
is_superset(x, y)
}
\arguments{
\item{x}{A numeric, character, or logical vector.}

\item{y}{A numeric, character, or logical vector.}
}
\value{
A logical value indicating whether \code{x} is a superset of \code{y}.
}
\description{
Determines if all elements in set \code{y} are also present in set \code{x}.
}
