% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{setopts_are_equal}
\alias{setopts_are_equal}
\title{Compare Sets for Equality}
\usage{
setopts_are_equal(x, y)
}
\arguments{
\item{x}{A vector of elements.}

\item{y}{A vector of elements.}
}
\value{
A logical value indicating whether the sets are equal (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Determine if the two sets are equal.
}
\examples{
if(interactive()){
setopts_are_equal(c(1, 2, 3), c(1, 2, 3))
# Output: TRUE
setopts_are_equal(c(1, 2, 3), c(3, 2, 1))
# Output: TRUE
setopts_are_equal(c(1, 2, 3), c(1, 2, 3, 4))
# Output: FALSE
}
}
