% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{setopts_common_elements}
\alias{setopts_common_elements}
\title{Find Common Elements}
\usage{
setopts_common_elements(x, y)
}
\arguments{
\item{x}{A vector of elements.}

\item{y}{A vector of elements.}
}
\value{
A vector of elements that are present in both sets.
}
\description{
Find the elements that are present in both sets.
}
\examples{
assertions:::setopts_common_elements(c(1, 2, 3), c(2, 3, 4))
# Output: [1] 2 3
}
