% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{setopts_count_exlusive_to_first}
\alias{setopts_count_exlusive_to_first}
\title{Count of Elements Exclusive to First Set}
\usage{
setopts_count_exlusive_to_first(x, y)
}
\arguments{
\item{x}{A vector of elements.}

\item{y}{A vector of elements.}
}
\value{
A scalar representing the number of elements that are in the first set but not in the second set.
}
\description{
Counts the number of elements that are in the first set but not in the second set.
}
\examples{
assertions:::setopts_count_exlusive_to_first(c(1, 2, 3), c(3, 4, 5))
# Output: 2
}
