% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{util_count_duplicates}
\alias{util_count_duplicates}
\title{Count the number of duplicated values in a vector}
\usage{
util_count_duplicates(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of duplicated values in the input vector.
}
\description{
This function returns the number of duplicated values in the input vector.
}
\examples{
if(interactive()) {
util_count_duplicates(c(1, 2, 2))  # returns 1
util_count_duplicates(c(1, 2, 3))  # returns 0
}
}
