% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{util_count_missing}
\alias{util_count_missing}
\title{Count the number of missing values in a vector}
\usage{
util_count_missing(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of missing values in the input vector.
}
\description{
This function returns the number of missing values (\code{NA}) in the input vector.
}
\examples{
assertions:::util_count_missing(c(1, 2, 3))  # returns 0
assertions:::util_count_missing(c(1, NA, 2))  # returns 1
}
