% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{util_get_duplicated_values}
\alias{util_get_duplicated_values}
\title{Get the duplicated values in a vector}
\usage{
util_get_duplicated_values(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of the duplicated values in the input vector.
}
\description{
This function returns a vector of the duplicated values in the input vector.
}
\examples{
if(interactive()) {
util_get_duplicated_values(c(1, 2, 2))  # returns 2
util_get_duplicated_values(c(1, 2, 3))  # returns NULL
}
}
