% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_files.R
\name{assert_file_exists}
\alias{assert_file_exists}
\title{Assert a file exists}
\usage{
assert_file_exists(x, msg = NULL, call = rlang::caller_env(), arg_name = NULL)
}
\arguments{
\item{x}{An object}

\item{msg}{A character string containing the error message if file \code{x} is does not exist}

\item{call}{Only relevant when pooling assert_ions into multi-assertion helper functions. See \link[cli]{cli_abort} for details.}

\item{arg_name}{Advanced use only. Name of the argument passed (default: NULL, will automatically extract arg_name).}
}
\value{
invisible(TRUE) if file \code{x} exists, otherwise aborts with the error message specified by \code{msg}
}
\description{
Assert that all files exist
}
\examples{
try({
assert_file_exists(system.file("package = assertions"))
assert_file_exists("foo") # Throws Error
})

}
\concept{assert_file}
