% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_testing.R
\name{assertion_names}
\alias{assertion_names}
\title{List assertion names}
\usage{
assertion_names(exclude_create_and_chain = TRUE)
}
\arguments{
\item{exclude_create_and_chain}{exclude assert_create and assert_create_chain (flag)}
}
\value{
unique set of assertion names (character)
}
\description{
List all assertion names
}
