% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_character_vector}
\alias{is_character_vector}
\title{Check if an object is a character vector}
\usage{
is_character_vector(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether \code{x} is a character vector.
}
\description{
Check if an object is a character vector
}
\concept{is_type}
