% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_vector}
\alias{is_vector}
\title{Check if an object is a vector
This function checks if an object is a vector}
\usage{
is_vector(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
A logical indicating whether \code{x} is a vector
}
\description{
Check if an object is a vector
This function checks if an object is a vector
}
\examples{
if(interactive()){
is_vector(1:10)
is_vector(list(1:10)) # Fails
}

}
\concept{is_type}
